big academy update
Courses have now a deadline date. It’s set to the startdate of the
first coursedate. Used to determine which courses are displayed in the
frontend.
Attendees have now a „presen“ field. Helpful in administration.
Booking courses now send an email (hurray!)
Booking courses now send a notification mail to administrators
attendee formtype in backend is now separate to frontend
attendee formtypes reduced and notes are now html5 formvalidation free
(not required)